cleanup: reorganize component structure
