+++ components/ResultsPage/EligibilityDetails.tsx
