+++ components/ResultsPage/ResultsTable.tsx
